Fixes #7247, move default Formatter class to new Numeric metrics formatter and use in Graph visualization since jqplot requires numeric values in JSON data.
