+++ plugins/CoreAdminHome/Controller.php
@@ -292,7 +292,7 @@ class Controller extends ControllerAdmin
         $view->idSite = Common::getRequestVar('idSite', $defaultIdSite, 'int');
 
         $view->defaultReportSiteName = Site::getNameFor($view->idSite);
-        $view->defaultSiteRevenue = \Piwik\Metrics\Formatter::getCurrencySymbol($view->idSite);
+        $view->defaultSiteRevenue = \Piwik\Metrics\Formatter\Api::getCurrencySymbol($view->idSite);
         $view->maxCustomVariables = CustomVariables::getMaxCustomVariables();
 
         $allUrls = APISitesManager::getInstance()->getSiteUrlsFromId($view->idSite);
