+++ plugins/CoreVisualizations/Visualizations/Graph.php
@@ -10,6 +10,7 @@ namespace Piwik\Plugins\CoreVisualizations\Visualizations;
 
 use Piwik\DataTable;
 use Piwik\DataTable\Row;
+use Piwik\Metrics\Formatter\Numeric;
 use Piwik\Piwik;
 use Piwik\Plugin\Visualization;
 
@@ -60,6 +61,8 @@ abstract class Graph extends Visualization
 
         $this->requestConfig->request_parameters_to_modify['disable_queued_filters'] = 1;
         $this->requestConfig->request_parameters_to_modify['format_metrics'] = 1;
+
+        $this->metricsFormatter = new Numeric();
     }
 
     /**
